// src/core/api/generated/spring/schemas/SystemConfigDTOSystemConfigType.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type SystemConfigDTOSystemConfigType = typeof SystemConfigDTOSystemConfigType[keyof typeof SystemConfigDTOSystemConfigType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SystemConfigDTOSystemConfigType = {
  PRODUCT: 'PRODUCT',
  INVENTORY: 'INVENTORY',
  USER: 'USER',
  CUSTOM: 'CUSTOM',
} as const;